separated orderBike into components and cleaned up orderBike and orderComponents
