+++ bike-erp/src/components/OrderService/OrderBike/BikeBillingStyles.ts
@@ -1,0 +1,14 @@
+import { makeStyles } from "@material-ui/core";
+
+const useStyles = makeStyles((theme) => ({
+  billingBox: {
+    margin: 20
+  },
+  item: {
+    float:"left",
+    left:0,
+    color: "red",
+  }
+}));
+
+export default useStyles;
