+++ bike-erp/src/components/OrderService/OrderBike/BikeModelStyles.ts
@@ -1,0 +1,13 @@
+import { makeStyles } from "@material-ui/core";
+
+const useStyles = makeStyles((theme) => ({
+  image: {
+    width: 500,
+    height: 305,
+  },
+  location: {
+    minWidth: 200,
+  }
+}));
+
+export default useStyles;
