--- bike-erp/src/components/OrderService/OrderBiling/OrderBiling.css
@@ -1,32 +1,0 @@
-.orderBiling {
-  padding: 10px;
-  height: 750px;
-  position: relative;
-
-}
-
-.total {
-  display: flex;
-  padding-top: 50px;
-}
-
-.item {
-  float:left;
-  left:0;
-  color: red;
-}
-
-.total {
-  padding-top: 50px;
-  position: absolute;
-  bottom: 0;
-}
-
-.contents {
-  overflow-y:auto;
-  height: 650px;
-}
-
-.billingBox {
-  margin: 20px;
-}
