+++ bike-erp/src/components/OrderService/OrderComponent/ComponentView.tsx
@@ -1,43 +1,27 @@
-import { CardMedia, Card, CardActions, Select, FormControl, InputLabel, MenuItem } from "@material-ui/core";
 import bike_logo from "../../../assets/images/login_bike_logo.png";
-import { Box,Typography } from "@material-ui/core";
-import "./ModelView.css";
+
+import { Box, CardMedia, Card, CardActions, Select, FormControl, InputLabel, MenuItem, Typography } from "@material-ui/core";
+import useStyles from "./ComponentViewStyles";
 
 const ModelView = ({ setSelectedLocation }: any) => {
+
+  const styles = useStyles();
 
   const onSelect = (event: React.ChangeEvent<{ value: unknown }>) => {
     setSelectedLocation(event.target.value as string)
   }
-  const styles = {
-    media: {
-      height: 0,
-      paddingTop: '60%',
-      marginTop: '30'
-    },
-    location: {
-      minWidth: 200,
-    },
-    image: {
-      width: "100%",
-      height:"100%"
-    }
-  };
+  
   return (
     <Box>
       <Card>
         <h2>MODEL</h2>
         <CardMedia title="bike-logo">
-          <img src={bike_logo} style={styles.image} alt="bike-logo" />
-          {
-            /* Dynamically create components from list */
-          }
+          <img src={bike_logo} className={styles.image} alt="bike-logo" />
         </CardMedia>
         <CardActions >
-          <FormControl style={styles.location}>
+          <FormControl className={styles.location}>
             <InputLabel>Location</InputLabel>
-            <Select name="componentLocation" id="compLoc" defaultValue={"None"}
-              onChange={onSelect}
-            >
+            <Select name="componentLocation" id="compLoc" defaultValue={"None"} onChange={onSelect}>
               <MenuItem selected disabled value={"None"}>
                 <em>None</em>
               </MenuItem>
@@ -48,7 +32,7 @@ const ModelView = ({ setSelectedLocation }: any) => {
           </FormControl>
         </CardActions>
       </Card>
-      <Typography className="message" variant="body1">*Please select a location, size and component before adding.</Typography>
+      <Typography className={styles.message} variant="body1">*Please select a location, size and component before adding.</Typography>
     </Box>
   );
 }
