+++ bike-erp/src/components/OrderService/OrderComponent/ComponentViewStyles.ts
@@ -1,0 +1,16 @@
+import { makeStyles } from "@material-ui/core";
+
+const useStyles = makeStyles((theme) => ({
+  location: {
+    minWidth: 200,
+  },
+  image: {
+    width: "100%",
+    height:"100%"
+  },
+  message: {
+    float: "left"
+  }
+}));
+
+export default useStyles;
