--- bike-erp/src/components/OrderService/OrderComponent/ModelView.css
@@ -1,29 +1,0 @@
-.componentView {
-  max-width: 800px;
-  object-fit: contain;
-  border: 1px solid black;
-}
-
-.model {
-  width: 500px;
-  height: 500px;
-  box-sizing: border-box;
-}
-
-.model img {
-  object-fit: contain;
-  width: 500px;
-  top: 50%;
-}
-
-.location .componentLocation, .location option {
-  background-color: white;
-}
-
-.location option {
-  padding: 5px;
-}
-
-.message {
-  float: left;
-}
