+++ bike-erp/src/pages/OrderService/OrderComponent/OrderComponent.tsx
@@ -1,14 +1,16 @@
-import { Typography } from "@material-ui/core";
-import Grid from "@material-ui/core/Grid";
 import { useState } from "react";
-import OrderBiling from "../../../components/OrderService/OrderBiling/OrderBiling";
-import Components from "../../../components/OrderService/Components";
-import ModelView from "../../../components/OrderService/OrderComponent/ModelView";
+
+import OrderBiling from "../../../components/OrderService/OrderComponent/ComponentBilling";
+import Components from "../../../components/OrderService/OrderComponent/Components";
+import ModelView from "../../../components/OrderService/OrderComponent/ComponentView";
+
+import { Grid, Typography } from "@material-ui/core";
 import useStyles from "./OrderComponentStyle";
 
 const OrderComponent = () => {
-  const [selectedLocation, setSelectedLocation] = useState("None");
+
   const styles = useStyles();
+  const [selectedLocation, setSelectedLocation] = useState("None");
 
   return (
     <div>
